@@ -4,7 +4,7 @@ version: 0.3.2
 homepage: https://github.com/RainwayApp/flutter_gamepad
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.14.0 <3.0.0"
 
 dependencies:
   flutter:
