@@ -104,7 +104,7 @@ class GamepadConnectedEvent implements GamepadEvent {
   final GamepadInfo gamepadInfo;
 
   const GamepadConnectedEvent(
-      {@required this.gamepadId, @required this.gamepadInfo});
+      {required this.gamepadId, required this.gamepadInfo});
 }
 
 /// A gamepad event that fires when a controller is disconnected.
@@ -119,7 +119,7 @@ class GamepadDisconnectedEvent implements GamepadEvent {
   final GamepadInfo gamepadInfo;
 
   const GamepadDisconnectedEvent(
-      {@required this.gamepadId, @required this.gamepadInfo});
+      {required this.gamepadId, required this.gamepadInfo});
 }
 
 /// A gamepad event describing a button state change.
@@ -138,10 +138,10 @@ class GamepadButtonEvent implements GamepadEvent {
   final bool pressed;
 
   const GamepadButtonEvent({
-    @required this.gamepadId,
-    @required this.button,
-    @required this.value,
-    @required this.pressed,
+    required this.gamepadId,
+    required this.button,
+    required this.value,
+    required this.pressed,
   });
 }
 
@@ -162,10 +162,10 @@ class GamepadThumbstickEvent implements GamepadEvent {
   final double y;
 
   const GamepadThumbstickEvent({
-    @required this.gamepadId,
-    @required this.thumbstick,
-    @required this.x,
-    @required this.y,
+    required this.gamepadId,
+    required this.thumbstick,
+    required this.x,
+    required this.y,
   });
 }
 
@@ -177,7 +177,7 @@ class GamepadButtonDebugEvent implements GamepadEvent {
   final String info;
 
   const GamepadButtonDebugEvent({
-    @required this.gamepadId,
-    @required this.info,
+    required this.gamepadId,
+    required this.info,
   });
 }
